+++ packages/kb-client/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @kb/client
 
+## 1.3.23
+
+### Patch Changes
+
+- Updated dependencies
+  - @kb/core@1.6.2
+
 ## 1.3.22
 
 ### Patch Changes